refactor - divide into mcp server secvices and mcp client services - remove gpt because it is no longer maintained
