+++ tabs/llm/llm_agents/gemini/gemini_mcp_client.py
@@ -7,7 +7,7 @@ from mcp import ClientSession
 from google.genai import types
 from google.genai.types import GenerateContentResponse
 
-from services.resume_saver_service import ResumeSaverService 
+from tabs.llm.llm_agents.agent_services.resume_saver_service import ResumeSaverService 
 
 
 class SmartMCPClient:
